[BE] refactor : delete 'lombok' annotation (@getter, @setter)
