+++ BE/side-dish-1/src/main/java/com/codesquad/sidedish1/presentation/Category.java
@@ -1,13 +1,10 @@
-package com.example.demo.Entity;
+package com.codesquad.sidedish1.presentation;
 
-import lombok.Getter;
-import lombok.Setter;
 import org.springframework.data.annotation.Id;
 
 import java.util.ArrayList;
 import java.util.List;
 
-@Getter @Setter
 public class Category {
 
     @Id
